Add support for custom redirect status codes
Status codes can now be configured for all rules, and additionally for individual rules.

The implementation doesn't add any meaningful overhead by default: any rules that use the set default status code (or 302 if none was set) have the same encoding as before. Custom codes are appended to the target after a single space. In that case, the only two additional operations on the request path are to check if the target is long enough (i.e. at least 5 characters) to have a status code in the first place, and then whether the 4th-to-last character is a space. If so, the code is extracted and parsed as a u16, otherwise the default code is used.